refactor(api): move dictToQueryString to existing utils file
